refactor, fix
