+++ guards.js
@@ -1,4 +1,16 @@
+const passport = require('passport');
+const JwtStrategy = require('passport-jwt').Strategy;
+const ExtractJwt = require('passport-jwt').ExtractJwt;
+const { jwtCallback } = require('./passport');
+const { jwt_secret } = require('./config');
 const db = require('./database');
+
+const options = {
+	jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
+	secretOrKey: jwt_secret,
+}
+passport.use(new JwtStrategy(options, jwtCallback));
+const auth = passport.authenticate('jwt', { session: false });
 
 const adminGuard = (req, res, next) => {
 	const user = db.getByEmail(req.user.email);
@@ -12,4 +24,6 @@ const adminGuard = (req, res, next) => {
 
 module.exports = {
 	adminGuard,
+	auth,
+	passport
 }
