+++ server.js
@@ -1,11 +1,12 @@
-require('dotenv').config();
-const PORT = process.env.PORT;
-
+const { port } = require('./config');
 const express = require('express');
 const cors = require('cors');
-const passport = require('./routes/auth').passport
+const { passport } = require('./guards');
 
-const authRouter = require('./routes/auth').router;
+const registerRouter = require('./routes/register');
+const loginRouter = require('./routes/login');
+const usersRouter = require('./routes/users');
+const postsRouter = require('./routes/posts');
 const accountsRouter = require('./routes/accounts');
 const categoriesRouter = require('./routes/categories');
 
@@ -15,14 +16,14 @@ app.use(cors());
 app.use(express.json());
 app.use(express.urlencoded({ extended: false }));
 
-
-
-
-app.use('/auth', authRouter);
+app.use('/register', registerRouter);
+app.use('/login', loginRouter);
+app.use('/users', usersRouter);
+app.use('/posts', postsRouter);
 app.use('/accounts', accountsRouter);
 app.use('/categories', categoriesRouter);
 app.use(passport.initialize());
 
-app.listen(PORT, () => {
-	console.log(`Server is listening on http://localhost:${PORT}`);
+app.listen(port, () => {
+	console.log(`Server is listening on http://localhost:${port}`);
 })
